Remove explicit multitenant mode in address controller
With this change the address controller no longer requires you to use
it in single or multitenant mode. If you deploy an address space to its
namespace, it will assume single tenant mode and not require cluster
admin privileges.

With this approach, you can choose to not deploy any address spaces by
default, in which case you are free to choose later if you want to run
in multitenant mode or not, and also gives you the ability to configure
properties of the single tenant address space.

This fixes #350