Alternate implementation of per-address settings
This creates an address-settings instance per plan (only for topic and queue plans with broker resource
requirement less than 1). The router is then configured to add prefixes to the name of the queue/topic,
with the prefix reflecting the plan so that it can be associated with the appropriate address-settings.