Create license metadata for agent module (#1904)
Uses license-reporter to generate a license.xml for both agent and the console.  These artefacts
are then merge into a single file that is included in the distribution.  license-reporter is also
used to create files containing the text of the license themselves. These too are include in the
distribution artefact.