Add IoT support to EnMasse via Eclipse Hono
This change adds initial support for IoT projects based on Eclipse Hono.
It also introduces golang support to the EnMasse projects.

The PR contains a new IoTProject CRD, an IoT operator, a custom garbage
collector, a tenant service based on the K8S CRD, a configurator for the
AMQP splitter, deployment scripts and a little bit of documentation.

This commit sums up the changes of the "iot" feature branch, and allow
further development in the "master" branch of EnMasse. So don't see this
as the complete feature, this is only just the beginning.

Co-Authored-By: Dejan Bosanac <dejan@nighttale.net>
Co-Authored-By: Ulf Lilleengen <lulf@pvv.ntnu.no>