Fix the PlanTest failing due to missing endpoint
This change waits after the update of the address space that the
resource version changes, only after that does it make sense to wait
to wait for the readiness of the address space.

What happened was that the update came, and immediately it was checked
that the address space is ready, which succeeded, but for the old state.
Then the test progressed, but actually ran into an "unready" address
space.