Add support for small sharded queues
Add another field 'partitions' to address plans, which will cause an
address to be allocated to multiple brokers. The broker credits will be
divided by the number of partitions, and the address will be created on
multiple brokers with a fraction of the credits on each.

Removes use of annotations for broker allocations as it is now stored in
status. Rewrite logic that depended on annotations to always use broker
status.