Handle the upgrade of pre-0.27 authentication services within the authentication service operator. (#2857)
* Handle the upgrade of pre-0.27 authentication services within the authentication service operator.

This extracts resources/claim details from the existing keycloak deployment and creates the appropiate authenticationservice record.  On OpenShift, if a postgresql
deploymentconfig exists and the database details tally with those of keycloak deployment a postgresql type authenticationservice record is created instead.

Ansible playbook now skips the creation of the authenticationservice if the keycloak deployment already exists.