Fix upgrade test (#3256)
* Create listener on old probe port to fix upgrade

This fixes an issue when upgrading where the router statefulset is not
completely rolled out after the upgrade. Once upgraded, the new agent
image will delete the old probe listener on the old router, causing the
readiness probe to fail. While the readiness probe is failing, the
rollout fails to proceed.

This fix re-adds the old probe which have to be kept around until
upgrading from 0.26.x is no longer supported.

* Re-add broker.xml for sharded-queue required to handle upgrade

This caused upgraded shared queue brokers to end up without a proper
broker.xml and thereby failed to drain as it was never connected to the
routers

* Add explanation of why file is needed