Check router mesh connectivity
This adds a check for router mesh connectivity by querying the routers
for its connected nodes in the router network. If the list of nodes does
not contain the number of ready router pods, the address space status is
set to false, and an error message is added to the status and in the
log.

Fixes #3355