Generate template assembly archive
This also changes the name of the assembly bundle from enmasse-${TAG} to
enmasse-${VERSION}, which will be different for master (i.e pom version
instead of tag which defaults to 'latest'). CI environments need to
accomodate this change.

Generating a dist zip will allow deploying the template as a maven
artifact.