JSON path fix - connection query
The JSON path queries can't support the traversal from one object to another (i.e. connection to address space) in the
same way you can express in GraphQL.  I've expressed the connnection queries to avoid this and modfied the mock to
give it the same behaviour of the actual implementation.