Let the tenant cleaner upload, but not override the protobut schema
This change tries to fix the issue when the tenant cleaner tries to
clean up a tenant from an infinispan cluster, which never got
the protobuf schema uploaded by the device registry.

In change 4824dbb7ae73eda147fccf9cb0b19553ed8f6699 the tenant cleaner
was changes to never upload the protobuf schema. But this may cause
the problem, that when an IoTProject is being created and deleted
without the device registry properly uploading the schema, then the
tenant cleaner will fail.

While in production the argument could be, that the device registry
should be brought up online, before starting to clean up IoTProjects,
in testing this might be an issue, especially when the test environment
gets broken by that situation.

So with this change, when the device registry is configured to upload
the schema, it will overwrite it (as before). While the tenant cleaner
will only set the schema if it does not exists.