Implement address purge (#3797)
Agent now has separate channels, one for commands and one for the flow of address and connection updates. Console-server authenticates to the agent separately when the user requests an operation such as address purge using the access token of the logged on user.

Internally, the AgentCollector got generalised into an AgentDelegate that now has a factory method CommandDelegate which produces a command delegate for a given access token. It is this object that the rest of the resolvers use to cause the mutation operation to occur.