Ensure infra uuid is not lost on replace
This adds an extra annotation with the namespace of the address space to the 'messaging' Service resource that is created for every address space. In the event that an address space is replaced and the annotations are lost, they can be recovered from the service resource in the reconciliation loop.

Fixes #3927