Allow configuring console to use impersonation
This extends the consoleservice configuration with settings to control
the use of impersonation of the console (disabled by default).

Propagate user to impersonate in xoauth2 username parameter for agent.

Use token reviews for authentication in agent

Also if impersonated return impersonated user as whoami

Split documentation into 2 guides for Kubernetes on how to configure
console either using impersonation or not using impersonation.

Fixes #4117