Add address canary check for routers and brokers (#4667)
* Add address canary check for routers and brokers

Add a periodic health check for standard address space, where
pre-configured canary addresses are added to the routers and
brokers. The periodic check will perform a send and receive to the
canary addresses, and report prometheus metrics with the result.

An example alert is provided that alerts on failed canaries for more
than 5 minutes.