Gracefully handle missing Prometheus CRDs in some cases (#4707)
If monitoring is disabled, the IoT reconciler tries to "delete" the
prometheus CRs. However that will fail if Prometheus CRDs are not
installed.

It would be possible to add another flag or check, however I think the
"ENABLE_MONTORING" env-var should already be enough. If monitoring
is enabled, the prometheus CRD must be installed, otherwise it is an
error. If monitoring is disabled, deleting the CR due to a missing
CRD can be ignored. If the user chooses to enable monitoring without
having the CRDs installed, an error is shown. It is still possible
to install prometheus CRDs afterwards. The reconciler will try again
and succeed the next time, as there are no stale caches. Using the
'HasApi' methods would cache the outcome, which is ok for system
resources, but wrong for CRDs.

In addition, the systemtests enable monitoring, although monitoring
is not installed. This is fixed as well.