Fix testing IoT on Minikube (#4757)
The final missing piece was the access token. Which you don't have
initially with Minikube. So now a service account, with an appropriate
role and binding will be created and the token used instead of the
default Kubernetes client API token.

This approach should properly work on all types of Kubernetes.