Modify PVC instead of statefulset claim template (#4754)
* Make sure storage does not change if not intended

* Modify PVC instead of statefulset claim template

This works around a limitation in Kubernetes where the statefulset
volume claim template cannot be modified (i.e. when resizing a
volume). The previous workaround was to use replace instead of patch,
which has now stopped working. The new workaround will ensure that the
existing claim template will stay as is, and instead the persistent
volume claim that was generated from the template previously, is now modified.

* Add changelog entry

* Shrinking volumes is not always supported

* Run messaging client before and after changing infra

* Use externalclient

* Add credentials

* Use durable messages