Purge address by .spec.address instead of .metadata.name
The agent uses the .spec.address as the key in its internal map, but the
purge request uses the .metadata.name as the key when doing the lookup.
This causes purge to fail if .metadata.name and .spec.address differ.

This fix changes the console server to pass the .spec.address instead of
.metadata.name and also using this as the lookup in agent.

An alternative would be to change the mapping internally to use
.metadata.name as the key, but this would complicate other uses of the
map that assumes it is keyed by .spec.address (such as router stats
retrieval).

Fixes #5195