Fix #5246: Avoid reliance of deprecated service-ca.crt (OpenShift)
On OpenShift 4, the operator now creates a special cm (with annotation) that causes OpenShift
to inject the service ca.  This service ca is passed to the console, standard auth and used
for messaging user controller.  In the case of the console/standard auth, the cm is mounted
onto the pod.  For the messaging user controller (part of the enmasse conttroller), the
controller awaits the projection of the config map.

On OpenShift 3.11, the original behaviour is retained.

Also on OpenShift 4 the beta version of the serving-cert-secret-name service annotation is
used for new services.  On OpenShift 3.11, the alpha version continues to be used.